fix #45: correct error message if @DataProvider is not static
The required validation and error message are already in the code.
And #validateInstanceMethods also already gathers this error. But
the also called #computeTestMethods() throws an Error because of already
found and added error. Such that the fix is to not call #computeTestMethods()
if there are already errors before. This just works by copying #validateInstanceMethods
and add this check.