migrate first generic subprojects configuration to Kotlin DSL (#123)
Note: Using a `String` for dependency configuration is ugly but the
alternative is to apply `java-library` plugins within `plugins` closure
which would also apply it to root project which is very much unwanted.
This also does not allow to use `java` closure and `javadoc` task.

Alternative: We could split build into multiple files with a lot
duplicated code, though.

Which one is better?